+++ release-notes.md
@@ -2,6 +2,7 @@
 
 ## 0.9.0
 - Added new `findAll()` method to [SerializedDataTypeRegistry](api/src/main/java/org/fuin/esc/api/SerializedDataTypeRegistry.java)
+- **Incompagible** The [ESGrpcEventStore](esgrpc/src/main/java/org/fuin/esc/esgrpc/ESGrpcEventStore.java) does no longer shutdown the `KurrentDBClient` to allow sharing the client between multiple instances as it is thread-safe.
 
 ## 0.8.0